+++ apple/Tests/TestCases/Profile/Collection/AccountsTests.swift
@@ -3,6 +3,7 @@ import Foundation
 import Sargon
 import SargonUniFFI
 import XCTest
+import SwiftyJSON
 
 final class AccountsTests: CollectionTest<Account> {
 	override class func sample() -> SUT {
@@ -23,5 +24,20 @@ final class AccountsTests: CollectionTest<Account> {
 		b.displayName = "Diff name, also crash" // different value on the element does not affect duplicates check, since it is ID based
 		profile.networks[0].accounts = [a, b] // Duplicates (by ID), not allowed => crash
 		let _ = profile.jsonData() // should crash
+	}
+
+	func test_json_decoding_of_profile_fails_if_accounts_contains_duplicates() throws {
+		var json = JSON(Profile.sample)
+		json["profileNetworks.accounts"] = [Account.sample, Account.sample]
+		XCTAssertThrowsError(try Profile(jsonData: json.rawData()))
+	}
+	
+	func test_json_decoding_of_profile_fails_if_accounts_contains_duplicated_ids() throws {
+		var json = JSON(Profile.sample)
+		let a = Account.sample
+		var b = a
+		b.displayName = "Diff name, also crash" // different value on the element does not affect duplicates check, since it is ID based
+		json["profileNetworks.accounts"] = [a, b]
+		XCTAssertThrowsError(try Profile(jsonData: json.rawData()))
 	}
 }
