+++ apple/Tests/Utils/FactorSourceTest.swift
@@ -4,18 +4,6 @@ import Sargon
 import SargonUniFFI
 import XCTest
 
-class SpecificFactorSourceTest<SUT_: FactorSourceProtocol>: FactorSourceTest<SUT_> {
-	
-	func test_extract() throws {
-		func doTest(_ sut: SUT) throws {
-			let embedded = sut.asGeneral
-			let extracted: SUT = try embedded.extract()
-			XCTAssertEqual(extracted, sut)
-		}
-		try SUT.sampleValues.forEach(doTest)
-	}
-}
-
 class FactorSourceTest<SUT_: BaseFactorSourceProtocol>: Test<SUT_> {
 	
 	func test_as_general_factorSourceID() {
